requirements.txt
torch
transformers
fastapi
uvicorn[standard]
python-multipart
Pillow
httpx
streamlit
